+++ src/app/pages/dashboard/dashboard.component.html
@@ -10,7 +10,7 @@
       >
     </li>
     <li>
-      <a class="govuk-link" [routerLink]="manualAccountCreationRoutes.accountDetails" id="manualAccountCreationLink"
+      <a class="govuk-link" [routerLink]="manualAccountCreationRoutes.createAccount" id="manualAccountCreationLink"
         >Manually Create Account</a
       >
     </li>
