+++ src/app/pages/manual-account-creation/account-details/account-details.component.spec.ts
@@ -1,17 +1,14 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { AccountDetailsComponent } from './account-details.component';
-import { ManualAccountCreationRoutes } from '@enums';
 import {
+  DEFENDANT_TYPES_STATE,
+  MANUAL_ACCOUNT_CREATION_PARENT_GUARDIAN_DETAILS_STATE,
   MANUAL_ACCOUNT_CREATION_ACCOUNT_DETAILS_STATE,
   MANUAL_ACCOUNT_CREATION_CONTACT_DETAILS_STATE,
   MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE,
-  MANUAL_ACCOUNT_CREATION_PARENT_GUARDIAN_DETAILS_STATE,
 } from '@constants';
 import { MacStateService } from '@services';
-import { IManualAccountCreationAccountDetailsState } from '@interfaces';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { BUSINESS_UNIT_REF_DATA_MOCK } from '@mocks';
+import { provideRouter } from '@angular/router';
 
 describe('AccountDetailsComponent', () => {
   let component: AccountDetailsComponent;
@@ -31,15 +28,13 @@ describe('AccountDetailsComponent', () => {
     };
 
     await TestBed.configureTestingModule({
-      imports: [AccountDetailsComponent, HttpClientTestingModule],
-      providers: [{ provide: MacStateService, useValue: mockMacStateService }],
+      imports: [AccountDetailsComponent],
+      providers: [{ provide: MacStateService, useValue: mockMacStateService }, provideRouter([])],
     }).compileComponents();
 
     fixture = TestBed.createComponent(AccountDetailsComponent);
     component = fixture.componentInstance;
-
-    component.macStateService.manualAccountCreation.accountDetails.businessUnit = null;
-
+    mockMacStateService = TestBed.inject(MacStateService);
     fixture.detectChanges();
   });
 
@@ -47,61 +42,33 @@ describe('AccountDetailsComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should handle form submission and navigate', () => {
+  it('should navigate on handleRoute', () => {
     const routerSpy = spyOn(component['router'], 'navigate');
-    const formData: IManualAccountCreationAccountDetailsState = {
-      businessUnit: 'Test',
-      defendantType: 'Test',
-    };
-
-    component.handleAccountDetailsSubmit(formData);
-
-    expect(mockMacStateService.manualAccountCreation.accountDetails).toEqual(formData);
-    expect(routerSpy).toHaveBeenCalledWith([ManualAccountCreationRoutes.createAccount]);
+    component.handleRoute('test');
+    expect(routerSpy).toHaveBeenCalledWith(['test']);
   });
 
-  it('should test handleUnsavedChanges', () => {
-    component.handleUnsavedChanges(true);
-    expect(component.macStateService.manualAccountCreation.unsavedChanges).toBeTruthy();
-    expect(component.stateUnsavedChanges).toBeTruthy();
+  it('should set defendantType correctly', () => {
+    mockMacStateService.manualAccountCreation.accountDetails.defendantType = 'adultOrYouthOnly';
 
-    component.handleUnsavedChanges(false);
-    expect(component.macStateService.manualAccountCreation.unsavedChanges).toBeFalsy();
-    expect(component.stateUnsavedChanges).toBeFalsy();
-  });
+    component['setDefendantType']();
 
-  it('should set the business unit for account details when there is only one business unit available and the current business unit is null', () => {
-    const response = { count: 1, refData: [BUSINESS_UNIT_REF_DATA_MOCK.refData[0]] };
-
-    component['setBusinessUnit'](response);
-
-    expect(component.macStateService.manualAccountCreation.accountDetails.businessUnit).toEqual(
-      BUSINESS_UNIT_REF_DATA_MOCK.refData[0].businessUnitName,
+    expect(component.defendantType).toEqual(
+      DEFENDANT_TYPES_STATE[mockMacStateService.manualAccountCreation.accountDetails.defendantType],
     );
   });
 
-  it('should not set the business unit for account details when there is only one business unit available but the current business unit is not null', () => {
-    const response = { count: 1, refData: [BUSINESS_UNIT_REF_DATA_MOCK.refData[0]] };
-
-    component.macStateService.manualAccountCreation.accountDetails.businessUnit =
-      BUSINESS_UNIT_REF_DATA_MOCK.refData[1].businessUnitName;
-
-    fixture.detectChanges();
-
-    component['setBusinessUnit'](response);
-
-    expect(component.macStateService.manualAccountCreation.accountDetails.businessUnit).toEqual(
-      BUSINESS_UNIT_REF_DATA_MOCK.refData[1].businessUnitName,
-    );
+  it('should set defendantType to be empty', () => {
+    mockMacStateService.manualAccountCreation.accountDetails.defendantType = 'test';
+    component['setDefendantType']();
+    expect(component.defendantType).toBe('');
   });
 
-  it('should not set the business unit for account details when there are multiple business units available', () => {
-    const response = BUSINESS_UNIT_REF_DATA_MOCK;
+  it('should not set defendantType', () => {
+    component.defendantType = '';
+    mockMacStateService.manualAccountCreation.accountDetails.defendantType = null;
 
-    component.macStateService.manualAccountCreation.accountDetails.businessUnit = null;
-
-    component['setBusinessUnit'](response);
-
-    expect(component.macStateService.manualAccountCreation.accountDetails.businessUnit).toBeNull();
+    component['setDefendantType']();
+    expect(component.defendantType).toBe('');
   });
 });
