+++ src/app/pages/manual-account-creation/company-details/company-details.component.html
@@ -4,7 +4,7 @@
   <div class="govuk-grid-column-two-thirds">
     <app-govuk-button buttonId="submitForm" type="submit">Continue</app-govuk-button>
     <app-govuk-cancel-link
-      (linkClickEvent)="handleRoute(manualAccountCreationRoutes.createAccount)"
+      (linkClickEvent)="handleRoute(manualAccountCreationRoutes.accountDetails)"
     ></app-govuk-cancel-link>
   </div>
 </div>
