+++ src/app/pages/manual-account-creation/contact-details/contact-details.component.ts
@@ -24,7 +24,7 @@ export class ContactDetailsComponent extends FormParentBaseComponent {
       stateChanges: true,
     };
 
-    this.routerNavigate(ManualAccountCreationRoutes.createAccount);
+    this.routerNavigate(ManualAccountCreationRoutes.accountDetails);
   }
 
   /**
