+++ src/app/pages/manual-account-creation/employer-details/employer-details-form/employer-details-form.component.html
@@ -95,7 +95,7 @@
     <div class="govuk-grid-column-two-thirds">
       <app-govuk-button buttonId="submitForm" type="submit">Return to account details</app-govuk-button>
       <app-govuk-cancel-link
-        (linkClickEvent)="handleRoute(manualAccountCreationRoutes.createAccount)"
+        (linkClickEvent)="handleRoute(manualAccountCreationRoutes.accountDetails)"
       ></app-govuk-cancel-link>
     </div>
   </div>
