+++ src/app/pages/manual-account-creation/employer-details/employer-details.component.spec.ts
@@ -59,7 +59,7 @@ describe('EmployerDetailsComponent', () => {
     component.handleEmployerDetailsSubmit(formData);
 
     expect(mockMacStateService.manualAccountCreation.employerDetails).toEqual(formData);
-    expect(routerSpy).toHaveBeenCalledWith([ManualAccountCreationRoutes.createAccount]);
+    expect(routerSpy).toHaveBeenCalledWith([ManualAccountCreationRoutes.accountDetails]);
   });
 
   it('should test handleUnsavedChanges', () => {
