+++ src/app/pages/manual-account-creation/employer-details/employer-details.component.ts
@@ -26,7 +26,7 @@ export class EmployerDetailsComponent extends FormParentBaseComponent {
       stateChanges: true,
     };
 
-    this.routerNavigate(ManualAccountCreationRoutes.createAccount);
+    this.routerNavigate(ManualAccountCreationRoutes.accountDetails);
   }
 
   /**
