+++ src/app/pages/manual-account-creation/parent-guardian-details/parent-guardian-details-form/parent-guardian-details-form.component.html
@@ -62,7 +62,7 @@
     <div class="govuk-grid-column-two-thirds">
       <app-govuk-button buttonId="submitForm" type="submit">Return to account details</app-govuk-button>
       <app-govuk-cancel-link
-        (linkClickEvent)="handleRoute(manualAccountCreationRoutes.createAccount)"
+        (linkClickEvent)="handleRoute(manualAccountCreationRoutes.accountDetails)"
       ></app-govuk-cancel-link>
     </div>
   </div>
