+++ src/app/pages/manual-account-creation/parent-guardian-details/parent-guardian-details.component.ts
@@ -26,7 +26,7 @@ export class ParentGuardianDetailsComponent extends FormParentBaseComponent {
       stateChanges: true,
     };
 
-    this.routerNavigate(ManualAccountCreationRoutes.createAccount);
+    this.routerNavigate(ManualAccountCreationRoutes.accountDetails);
   }
 
   /**
